requirements.txt
streamlit==1.37.1
streamlit-webrtc==0.47.7
opencv-python-headless==4.10.0.84
av==12.0.0
aiortc==1.9.0
tensorflow-cpu==2.15.0
numpy==1.26.4